Fixed bug [#1423] snapshots.sh needs to check error when checking out cvs: deleted temp working dir after tgz has been generated for each project and added test on cvs command in case of pb with a given repository

+++ gforge-plugin-scmcvs/bin/snapshots.sh
@@ -30,10 +30,12 @@ case "$1" in
                 # Make tgz archive
 		mkdir -p $work_dir/$dir-scm-$today
 		cd $work_dir/$dir-scm-$today
-		cvs -f -Q -d :local:$CVSROOT/$dir co -P .
-		cd $work_dir
-		tar czf $dir-scm-latest.tar.gz $dir-scm-$today
-		mv $dir-scm-latest.tar.gz $SCMSNAPSHOTSDIR
+		if cvs -f -Q -d :local:$CVSROOT/$dir co -P .; then
+		    cd $work_dir
+		    tar czf $dir-scm-latest.tar.gz $dir-scm-$today
+		    mv $dir-scm-latest.tar.gz $SCMSNAPSHOTSDIR
+		    rm -fr $dir-scm-$today
+		fi
 	    fi
         done
 	;;
